refactor(rsc): move extra
