+++ packages/rsc/src/extra/browser.tsx
@@ -2,9 +2,9 @@ import { tinyassert } from "@hiogawa/utils";
 import React from "react";
 import ReactDomClient from "react-dom/client";
 import ReactClient from "react-server-dom-webpack/client.browser";
-import { setRequireModule } from "./core/client-browser";
+import { setRequireModule } from "../core/client-browser";
+import type { CallServerCallback } from "../types";
 import type { RscPayload } from "./server";
-import type { CallServerCallback } from "./types";
 import { getRscScript } from "./utils/rsc-script";
 
 export async function hydrate(options?: {
