+++ packages/rsc/src/extra/utils/rsc-script.ts
